chore: replace epic dataDownload
