+++ src/actions/dataDownload.js
@@ -10,16 +10,3 @@ export const openDataDownloadDialog = (layerid) => ({
 export const closeDataDownloadDialog = () => ({
     type: types.DATA_DOWNLOAD_CLOSE_DIALOG,
 })
-
-export const startDataDownload = (payload) => ({
-    type: types.DATA_DOWNLOAD_START,
-    payload,
-})
-
-export const cancelDataDownload = () => ({
-    type: types.DATA_DOWNLOAD_CANCEL,
-})
-
-export const dataDownloadSuccess = () => ({
-    type: types.DATA_DOWNLOAD_SUCCESS,
-})
