+++ src/components/layers/download/DataDownloadDialog.js
@@ -1,12 +1,9 @@
+import { useD2 } from '@dhis2/app-runtime-adapter-d2'
 import i18n from '@dhis2/d2-i18n'
 import { Modal, ModalTitle, ModalContent, ModalActions } from '@dhis2/ui'
-import PropTypes from 'prop-types'
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
-import {
-    closeDataDownloadDialog,
-    startDataDownload,
-} from '../../../actions/dataDownload.js'
+import React, { useState } from 'react'
+import { useSelector, useDispatch } from 'react-redux'
+import { closeDataDownloadDialog } from '../../../actions/dataDownload.js'
 import { EVENT_LAYER } from '../../../constants/layers.js'
 import {
     META_DATA_FORMAT_ID,
@@ -15,6 +12,7 @@ import {
 } from '../../../util/geojson.js'
 import DataDownloadDialogActions from './DataDownloadDialogActions.js'
 import DataDownloadDialogContent from './DataDownloadDialogContent.js'
+import { downloadData } from './downloadData.js'
 
 const formatOptionsFlat = [
     META_DATA_FORMAT_ID,
@@ -26,104 +24,88 @@ const formatOptions = formatOptionsFlat.map((name, i) => ({
     name,
 }))
 
-class DataDownloadDialog extends Component {
-    static propTypes = {
-        closeDialog: PropTypes.func.isRequired,
-        downloading: PropTypes.bool.isRequired,
-        open: PropTypes.bool.isRequired,
-        startDownload: PropTypes.func.isRequired,
-        aggregations: PropTypes.object,
-        error: PropTypes.string,
-        layer: PropTypes.object,
+const DataDownloadDialog = () => {
+    const { d2 } = useD2()
+    const [selectedFormat, setSelectedFormat] = useState(2)
+    const [humanReadable, setHumanReadable] = useState(true)
+    const [isDownloading, setIsDownloading] = useState(false)
+    const [error, setError] = useState(null)
+
+    const dispatch = useDispatch()
+    const { mapViews } = useSelector((state) => state.map)
+    const aggregations = useSelector((state) => state.aggregations)
+    const { layerid, dialogOpen } = useSelector((state) => state.dataDownload)
+
+    const layer = mapViews.find((l) => l.id === layerid)
+
+    const onChangeFormat = (format) => {
+        setError(null)
+        setSelectedFormat(format.id - 1)
     }
 
-    state = {
-        selectedFormatOption: 2,
-        humanReadableChecked: true,
+    const onChangeHumanReadable = (isChecked) => {
+        setError(null)
+        setHumanReadable(isChecked)
     }
 
-    onChangeFormatOption = (newValue) => {
-        this.setState({ selectedFormatOption: newValue.id - 1 })
-    }
-    onCheckHumanReadable = (isChecked) => {
-        this.setState({ humanReadableChecked: isChecked })
+    const onCloseDialog = () => {
+        setError(null)
+        dispatch(closeDataDownloadDialog())
     }
 
-    onStartDownload = () => {
-        const { layer, aggregations } = this.props
-        const {
-            selectedFormatOption,
-            humanReadableChecked: humanReadableKeys,
-        } = this.state
-        const format = formatOptionsFlat[selectedFormatOption]
-
-        this.props.startDownload({
-            layer,
-            aggregations,
-            format,
-            humanReadableKeys,
-        })
-    }
-
-    render() {
-        const { open, layer, downloading, error, closeDialog } = this.props
-
-        if (!open || !layer) {
-            return null
+    const onStartDownload = async () => {
+        setIsDownloading(true)
+        setError(null)
+        try {
+            await downloadData({
+                layer,
+                aggregations: aggregations[layerid],
+                format: formatOptionsFlat[selectedFormat],
+                humanReadableKeys: humanReadable,
+                d2,
+            })
+            setIsDownloading(false)
+            onCloseDialog()
+        } catch (e) {
+            setIsDownloading(false)
+            setError(e.message)
         }
-
-        const layerType = layer.layer,
-            isEventLayer = layerType === EVENT_LAYER
-
-        return (
-            <Modal position="middle" onClose={closeDialog}>
-                <ModalTitle>{i18n.t('Download Layer Data')}</ModalTitle>
-                <ModalContent>
-                    <DataDownloadDialogContent
-                        isEventLayer={isEventLayer}
-                        error={error}
-                        layerName={layer.name}
-                        formatOptions={formatOptions}
-                        selectedFormatOption={
-                            this.state.selectedFormatOption + 1
-                        }
-                        humanReadableChecked={this.state.humanReadableChecked}
-                        onChangeFormatOption={this.onChangeFormatOption}
-                        onCheckHumanReadable={this.onCheckHumanReadable}
-                    />
-                </ModalContent>
-                <ModalActions>
-                    <DataDownloadDialogActions
-                        downloading={downloading}
-                        onStartClick={this.onStartDownload}
-                        onCancelClick={closeDialog}
-                    />
-                </ModalActions>
-            </Modal>
-        )
     }
-}
 
-const mapStateToProps = ({ dataDownload, map, aggregations = {} }) => {
-    const { layerid, dialogOpen: open, downloading, error } = dataDownload
-    const layer = map.mapViews.find((l) => l.id === layerid)
+    if (!dialogOpen || !layer) {
+        return null
+    }
 
-    if (open && !layer) {
+    if (dialogOpen && !layer) {
         console.error(
             'Tried to open data download dialog without specifying a source layer!'
         )
     }
 
-    return {
-        open,
-        layer,
-        downloading,
-        error,
-        aggregations: aggregations[layerid],
-    }
+    return (
+        <Modal position="middle" onClose={onCloseDialog}>
+            <ModalTitle>{i18n.t('Download Layer Data')}</ModalTitle>
+            <ModalContent>
+                <DataDownloadDialogContent
+                    isEventLayer={layer.layer === EVENT_LAYER}
+                    error={error}
+                    layerName={layer.name}
+                    formatOptions={formatOptions}
+                    selectedFormatOption={selectedFormat + 1}
+                    humanReadableChecked={humanReadable}
+                    onChangeFormatOption={onChangeFormat}
+                    onCheckHumanReadable={onChangeHumanReadable}
+                />
+            </ModalContent>
+            <ModalActions>
+                <DataDownloadDialogActions
+                    downloading={isDownloading}
+                    onStartClick={onStartDownload}
+                    onCancelClick={onCloseDialog}
+                />
+            </ModalActions>
+        </Modal>
+    )
 }
 
-export default connect(mapStateToProps, {
-    closeDialog: closeDataDownloadDialog,
-    startDownload: startDataDownload,
-})(DataDownloadDialog)
+export default DataDownloadDialog
