+++ src/components/map/layers/EventLayer.js
@@ -1,9 +1,9 @@
 import { getInstance as getD2 } from 'd2'
 import React from 'react'
 import { EVENT_COLOR, EVENT_RADIUS } from '../../../constants/layers.js'
-import { getAnalyticsRequest } from '../../../loaders/eventLoader.js'
 import { getContrastColor } from '../../../util/colors.js'
 import { filterData } from '../../../util/filter.js'
+import { getAnalyticsRequest } from '../../../util/getAnalyticsRequest.js'
 import { getDisplayPropertyUrl } from '../../../util/helpers.js'
 import { formatCount } from '../../../util/numbers.js'
 import EventPopup from './EventPopup.js'
@@ -84,7 +84,8 @@ class EventLayer extends Layer {
                     d2 = d2 || (await getD2())
 
                     eventRequest =
-                        eventRequest || (await getAnalyticsRequest(this.props))
+                        eventRequest ||
+                        (await getAnalyticsRequest(this.props, d2))
 
                     eventRequest = eventRequest
                         .withBbox(params.bbox)
