+++ src/epics/index.js
@@ -4,7 +4,6 @@ import 'rxjs/add/operator/filter'
 import 'rxjs/add/operator/do'
 import 'rxjs/add/operator/concatMapTo'
 import 'rxjs/add/observable/empty'
-import dataDownloadEpics from './dataDownload.js'
 import orgUnitEpis from './orgUnits.js'
 
 const errorEpic = (action$) =>
@@ -13,4 +12,4 @@ const errorEpic = (action$) =>
         .do((action) => console.error(action.error))
         .concatMapTo(Observable.empty()) // Avoid infinite loop, same as .map(action => Observable.empty()).concatAll()
 
-export default combineEpics(errorEpic, orgUnitEpis, dataDownloadEpics)
+export default combineEpics(errorEpic, orgUnitEpis)
