+++ src/loaders/eventLoader.js
@@ -7,20 +7,15 @@ import {
     EVENT_COLOR,
     EVENT_RADIUS,
 } from '../constants/layers.js'
-import { getEventColumns } from '../epics/dataDownload.js'
 import {
-    getOrgUnitsFromRows,
     getFiltersFromColumns,
     getFiltersAsText,
     getPeriodFromFilters,
     getPeriodNameFromId,
 } from '../util/analytics.js'
 import { cssColor, getContrastColor } from '../util/colors.js'
-import {
-    createEventFeatures,
-    addStyleDataItem,
-    getBounds,
-} from '../util/geojson.js'
+import { createEventFeatures, getBounds } from '../util/geojson.js'
+import { getAnalyticsRequest } from '../util/getAnalyticsRequest.js'
 import { styleByDataItem } from '../util/styleByDataItem.js'
 import { formatStartEndDate, getDateArray } from '../util/time.js'
 
@@ -80,7 +75,7 @@ const loadEventLayer = async (config) => {
 
     config.isExtended = showDataTable
 
-    const analyticsRequest = await getAnalyticsRequest(config)
+    const analyticsRequest = await getAnalyticsRequest(config, d2)
     let alert
 
     config.name = programStage.name
@@ -113,7 +108,8 @@ const loadEventLayer = async (config) => {
         config.outputIdScheme = 'ID' // Required for StyleByDataItem to work
         const { names, data, response } = await loadData(
             analyticsRequest.withPageSize(EVENT_CLIENT_PAGE_SIZE), // DHIS2-10742,
-            config
+            config,
+            d2
         )
         const { total } = response.metaData.pager
 
@@ -192,90 +188,12 @@ const loadEventLayer = async (config) => {
     }
 }
 
-// Also used to query for server cluster in map/EventLayer.js
-// TODO: Use DataIDScheme / OutputIDScheme instead of requesting all metaData (which can easily dwarf the actual response data)
-export const getAnalyticsRequest = async ({
-    program,
-    programStage,
-    filters,
-    startDate,
-    endDate,
-    rows,
-    columns,
-    styleDataItem,
-    eventStatus,
-    eventCoordinateField,
-    relativePeriodDate,
-    isExtended,
-}) => {
-    const orgUnits = getOrgUnitsFromRows(rows)
-    const period = getPeriodFromFilters(filters)
-    const dataItems = addStyleDataItem(
-        columns.filter(isValidDimension),
-        styleDataItem
-    )
-
-    // Add "display in reports" columns that are not already present
-    if (isExtended) {
-        const displayColumns = await getEventColumns({ programStage })
-
-        displayColumns.forEach((col) => {
-            if (!dataItems.find((item) => item.dimension === col.dimension)) {
-                dataItems.push(col)
-            }
-        })
-    }
-
-    const d2 = await getD2()
-
-    let analyticsRequest = new d2.analytics.request()
-        .withProgram(program.id)
-        .withStage(programStage.id)
-        .withCoordinatesOnly(true)
-
-    analyticsRequest = period
-        ? analyticsRequest.addPeriodFilter(period.id)
-        : analyticsRequest.withStartDate(startDate).withEndDate(endDate)
-
-    if (relativePeriodDate) {
-        analyticsRequest =
-            analyticsRequest.withRelativePeriodDate(relativePeriodDate)
-    }
-
-    analyticsRequest = analyticsRequest.addOrgUnitDimension(
-        orgUnits.map((ou) => ou.id)
-    )
-
-    if (dataItems) {
-        dataItems.forEach((item) => {
-            analyticsRequest = analyticsRequest.addDimension(
-                item.dimension,
-                item.filter
-            )
-        })
-    }
-
-    if (eventCoordinateField) {
-        // If coordinate field other than event coordinate
-        analyticsRequest = analyticsRequest
-            .addDimension(eventCoordinateField) // Used by analytics/events/query/
-            .withCoordinateField(eventCoordinateField) // Used by analytics/events/count and analytics/events/cluster
-    }
-
-    if (eventStatus && eventStatus !== 'ALL') {
-        analyticsRequest = analyticsRequest.withEventStatus(eventStatus)
-    }
-
-    return analyticsRequest
-}
-
 export const getCount = async (request) => {
     const d2 = await getD2()
     return await d2.analytics.events.getCount(request)
 }
 
-export const loadData = async (request, config = {}) => {
-    const d2 = await getD2()
+export const loadData = async (request, config = {}, d2) => {
     const response = await d2.analytics.events.getQuery(request)
 
     const { data, names } = createEventFeatures(response, config)
@@ -333,10 +251,5 @@ const getFilterOptionNames = async (filters, headers) => {
         {}
     )
 }
-
-// Empty filter sometimes returned for saved maps
-// Dimension without filter and empty items array returns false
-const isValidDimension = ({ dimension, filter, items }) =>
-    Boolean(dimension && (filter || !items || items.length))
 
 export default eventLoader
